requirements.txt
APScheduler==3.8.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Faker==9.3.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gevent==21.8.0
greenlet==1.1.2
grequests==0.6.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg==3.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
regex==2021.10.8
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
text-unidecode==1.3
tomli==1.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
<<<<<<< HEAD
urllib3==1.26.7
=======
>>>>>>> scheduler-branch
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
